refactor: 도메인 관심사에 따른 패키지 분리
