+++ src/main/java/domain/Player.java
@@ -1,5 +1,7 @@
 package domain;
 
+import domain.card.Cards;
+
 public class Player extends User {
 
 	public Player(final String name, Cards cards) {
