+++ src/main/java/domain/card/Suits.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.card;
 
 public enum Suits {
     HEART("하트"),
