+++ src/main/java/domain/strategy/ShuffleStrategy.java
@@ -1,6 +1,8 @@
-package domain;
+package domain.strategy;
 
 import java.util.List;
+
+import domain.card.Card;
 
 public interface ShuffleStrategy {
 	void shuffle(List<Card> cards);
