+++ src/test/java/domain/UsersTest.java
@@ -12,6 +12,12 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.MethodSource;
 
+import domain.card.Card;
+import domain.card.Deck;
+import domain.card.Denomination;
+import domain.card.Suits;
+import domain.strategy.RandomShuffleStrategy;
+
 public class UsersTest {
 
 	private Deck deck;
