+++ src/test/java/domain/card/DeckTest.java
@@ -1,9 +1,11 @@
-package domain;
+package domain.card;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+
+import domain.strategy.RandomShuffleStrategy;
 
 public class DeckTest {
